jobgate_appointment_system/backend/jobgate_appointment_system/urls.py
"""
URL configuration for jobgate_appointment_system project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
# jobgate_appointment_system/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from django.views.generic import RedirectView
from common.views import APIRootView
from django.http import JsonResponse
from dj_rest_auth.views import UserDetailsView # type: ignore
from django.contrib.auth import views as auth_views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

def health_check(request):
    """Simple health check endpoint"""
    return JsonResponse({
        'status': 'healthy',
        'message': 'JOBGATE API is running',
        'version': '1.0.0'
    })

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/health/', health_check, name='health-check'),
    path('api/', APIRootView.as_view(), name='api-root'), 
    path('api/users/', include('users.urls')),
    path('api/universities/', include('universities.urls')),
    path('api/appointments/', include('appointments.urls')),
    path('', RedirectView.as_view(url='/admin/', permanent=False)),

    # NEW: JWT routes
    path('api/auth/jwt/create/', TokenObtainPairView.as_view(), name='jwt-create'),
    path('api/auth/jwt/refresh/', TokenRefreshView.as_view(), name='jwt-refresh'),
    path('api/auth/', include('dj_rest_auth.urls')),         # login, logout, password, etc.
    path('api/auth/registration/', include('dj_rest_auth.registration.urls')),
    path('api/auth/users/me/', UserDetailsView.as_view(), name='rest_user_details'),
    path(
        'auth/reset-password/confirm/<uidb64>/<token>/',
        auth_views.PasswordResetConfirmView.as_view(),
        name='password_reset_confirm'          # <-- this is the missing name
    ),
]

if settings.DEBUG:
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
